+++ src/frontend/components/JoinDiscussionLeaveButton/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { defineMessages, FormattedMessage } from 'react-intl';
 import { withRouter } from 'react-router';